Updated the scene name pointer path.
The new value indicates the currently loading scene, and remains as the
active scene. This value is way more stable and prevents unhooking,
while working just as well as the previous one for detecting the split
condition. Very rarely this value seems corrupt, so some extra care is
needed surrounding state logic.